node: Prevent deadlocks due to blocked writes.
Because the BBS mostly uses a relatively simple
threading model (one thread per node, with
blocking I/O), write can block and previously
this has not been handled well. In particular,
a common deadlock observed is that if a node
is blocked on I/O, since the node locked is held,
any attempts to print the node list will deadlock
the entire BBS, since the node traversal locked
each node as it traversed. This is removed as
the per-node locking was not really necessary.

The more comprehensive fix here takes further
steps to ensure that the system as a whole
does not deadlock, even if individual node threads
block due to blocked writes (which is fine,
but should not be allowed to disrupt other
threads).

All node I/O to file descriptors now uses
bbs_node_fd_ functions, which properly serialize
writes. dprintf is no longer used to write to
nodes. If the node thread is doing the write,
the node is locked as usual since this thread
is allowed to block insofar as it is no disruptive
to the rest of the system. The more interesting
case is when a thread that is not handling a node
tries to write to it. This is done, for example,
in the IMAP and IRC servers when the action of
one client causes messages to be sent to other
clients. However, when writing, if that node is
already blocked, then the other thread would also
block, and this is very bad. To mitigate this, rather
than locking the node outright, we now use trylock
and abort if we cannot acquire the lock. Additionally,
non-blocking writes are used in this case to avoid
blocking indefinitely.

Because locking is now handled here, this also simplifies
application usage, since the IMAP and IRC modules no longer
need to provide their own locking to serialize writes to
node file descriptors, and as I/O is more concentrated
to a few core functions, there should be fewer places
that things can go wrong.

+++ include/readline.h
@@ -69,7 +69,7 @@ int readline_bytes_available(struct readline_data *restrict rldata, int process)
  *         not including the delimiter.
  * \note The actual number of bytes read may be greater than the number of bytes returned. These bytes will be returned in subsequent calls to this function.
  */
-int bbs_readline(int fd, struct readline_data *restrict rldata, const char *restrict delim, int timeout);
+ssize_t bbs_readline(int fd, struct readline_data *restrict rldata, const char *restrict delim, int timeout);
 
 /*!
  * \brief Read exactly n bytes from a file descriptor and write them to another file descriptor
@@ -82,7 +82,7 @@ int bbs_readline(int fd, struct readline_data *restrict rldata, const char *rest
  * \return number of bytes read
  * \note The written data is NOT NUL-terminated, this is a binary operation
  */
-int bbs_readline_getn(int fd, int destfd, struct readline_data *restrict rldata, int timeout, size_t n);
+ssize_t bbs_readline_getn(int fd, int destfd, struct readline_data *restrict rldata, int timeout, size_t n);
 
 /*!
  * \brief Read exactly n bytes from a file descriptor and append them to a dyn_str
@@ -94,7 +94,7 @@ int bbs_readline_getn(int fd, int destfd, struct readline_data *restrict rldata,
  * \retval -1 on failure
  * \return number of bytes read
  */
-int bbs_readline_getn_dynstr(int fd, struct dyn_str *restrict dynstr, struct readline_data *restrict rldata, int timeout, size_t n);
+ssize_t bbs_readline_getn_dynstr(int fd, struct dyn_str *restrict dynstr, struct readline_data *restrict rldata, int timeout, size_t n);
 
 /*!
  * \brief Read exactly n bytes from a file descriptor and store them in a dynamically allocated buffer
@@ -116,7 +116,7 @@ char *bbs_readline_getn_str(int fd, struct readline_data *restrict rldata, int t
  * \param n Number of bytes to read and discard
  * \retval Number of bytes discarded on success, -1 on failure
  */
-int bbs_readline_discard_n(int fd, struct readline_data *restrict rldata, int timeout, size_t n);
+ssize_t bbs_readline_discard_n(int fd, struct readline_data *restrict rldata, int timeout, size_t n);
 
 /*!
  * \brief Set the boundary until which data should be read
