+++ nets/net_imap/imap_client.h
@@ -75,7 +75,7 @@ struct imap_client *imap_client_get_by_url(struct imap_session *imap, const char
  * \retval -1 on failure
  * \return Same as write
  */
-int __attribute__ ((format (gnu_printf, 3, 4))) __imap_client_send_log(struct imap_client *client, int log, const char *fmt, ...);
+ssize_t __attribute__ ((format (gnu_printf, 3, 4))) __imap_client_send_log(struct imap_client *client, int log, const char *fmt, ...);
 
 #define imap_client_send(client, fmt, ...) __imap_client_send_log(client, 0, fmt, ## __VA_ARGS__)
 #define imap_client_send_log(client, fmt, ...) __imap_client_send_log(client, 1, fmt, ## __VA_ARGS__)
